@@ -4,22 +4,24 @@ publish_to: "Uhamka"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  auto_size_text: ^2.1.0
-  auto_size_text_field: ^0.1.7
+  auto_size_text: ^3.0.0-nullsafety.0
   cloud_firestore: ^2.5.3
   cupertino_icons: ^1.0.2
   curved_navigation_bar: ^1.0.1
+  external_app_launcher: ^2.0.1
   firebase_core: ^1.7.0
   flutter:
     sdk: flutter
   flutter_launcher_icons: ^0.9.2
   flutter_native_splash: ^1.2.4
   flutter_svg: ^0.23.0+1
+  http: ^0.13.4
   liquid_swipe: ^2.1.1
-  percent_indicator: ^2.1.7+2
+  percent_indicator: ^3.4.0
+  url_launcher: ^6.0.12
 
 dev_dependencies:
   flutter_test:
@@ -47,13 +49,19 @@ flutter:
   uses-material-design: true
 
   assets:
-    - assets/SplashScreen.png
+    - assets/App_SplashScreen.png
     - assets/AppIcon.png
     - assets/AppIcon_2.png
     - assets/OnBoarding_Illustration_1.svg
     - assets/OnBoarding_Illustration_2.svg
     - assets/OnBoarding_Illustration_3.svg
     - assets/OnBoarding_IconNext.svg
+    - assets/Icon_ChecklistSnackbar.svg
+    - assets/Icon_WarningSnackbar.svg
+    - assets/Icon_ErrorSnackbar.svg
+    - assets/Frame_DialogPrimary.svg
+    - assets/Icon_CloseDialog.svg
+    - assets/Illustration_SendMailSuccess.svg
 
   fonts:
     - family: Folks
