@@ -3,30 +3,33 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-SizedBox DefaultTextField({
+SizedBox DefaultTextFieldSingle({
   required String label,
   required TextEditingController controller,
   required TextInputType textInputType,
   required TextInputAction textInputAction,
   bool obscureText = false,
-  IconButton? iconSuffix,
   Function(String)? onChanged,
-  String? errorText,
 }) {
   return SizedBox(
     width: double.infinity,
     child: TextFormField(
+      textAlign: TextAlign.center,
       obscureText: obscureText,
       scrollPadding: EdgeInsets.all(0),
       controller: controller,
       decoration: InputDecoration(
         contentPadding: EdgeInsets.symmetric(vertical: 5, horizontal: 10),
-        labelText: label,
-        labelStyle: TextStyle(fontFamily: "Folks"),
-        errorStyle: TextStyle(fontFamily: "Folks"),
-        errorText: errorText,
+        label: SizedBox(
+          width: double.infinity,
+          child: Center(
+              child: Text(
+            label,
+            style: TextStyle(fontFamily: "Folks"),
+          )),
+        ),
+        floatingLabelBehavior: FloatingLabelBehavior.never,
         border: OutlineInputBorder(),
-        suffixIcon: iconSuffix,
       ),
       keyboardType: textInputType,
       textInputAction: textInputAction,
