@@ -1,6 +1,7 @@
+// ignore_for_file: non_constant_identifier_names
+
 import 'package:flutter/material.dart';
 
-// ignore: non_constant_identifier_names
 Flexible FlexSpace(int flex) {
   return Flexible(flex: flex, child: Container());
 }
